# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Gestion Salud API",
        "description": "Documentación de la API de Gestión de Salud",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/recetas-medicas": {
            "get": {
                "tags": [
                    "Recetas Médicas"
                ],
                "summary": "Listar recetas médicas",
                "operationId": "ac939d560f132a89dcb5fbc5144bf849",
                "parameters": [
                    {
                        "name": "cita_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "medicamento_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "paciente_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "medico_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "estado_despacho",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pendiente",
                                "parcial",
                                "despachada"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de recetas médicas"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Recetas Médicas"
                ],
                "summary": "Registrar receta médica",
                "operationId": "db89291c09abb24f6449731dd6d004c4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "cita_id",
                                    "medicamento_id"
                                ],
                                "properties": {
                                    "cita_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "medicamento_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "dosis": {
                                        "type": "string",
                                        "example": "500 mg",
                                        "nullable": true
                                    },
                                    "frecuencia": {
                                        "type": "string",
                                        "example": "Cada 8 horas",
                                        "nullable": true
                                    },
                                    "duracion": {
                                        "type": "string",
                                        "example": "7 dias",
                                        "nullable": true
                                    },
                                    "indicaciones": {
                                        "type": "string",
                                        "example": "Tomar despues de las comidas",
                                        "nullable": true
                                    },
                                    "estado_despacho": {
                                        "type": "string",
                                        "nullable": true,
                                        "enum": [
                                            "pendiente",
                                            "parcial",
                                            "despachada"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Receta médica creada"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "422": {
                        "description": "Datos inválidos"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/recetas-medicas/{id}": {
            "get": {
                "tags": [
                    "Recetas Médicas"
                ],
                "summary": "Detalle de receta médica",
                "operationId": "3d9b00657cb0ac9250e3cb0030eab445",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Receta médica encontrada"
                    },
                    "404": {
                        "description": "No encontrada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Recetas Médicas"
                ],
                "summary": "Actualizar receta médica",
                "operationId": "c3e4176b0be39474f129706cadb7f5a0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "cita_id": {
                                        "type": "integer"
                                    },
                                    "medicamento_id": {
                                        "type": "integer"
                                    },
                                    "dosis": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "frecuencia": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "duracion": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "indicaciones": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "estado_despacho": {
                                        "type": "string",
                                        "enum": [
                                            "pendiente",
                                            "parcial",
                                            "despachada"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Receta médica actualizada"
                    },
                    "404": {
                        "description": "No encontrada"
                    },
                    "422": {
                        "description": "Datos inválidos"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Recetas Médicas"
                ],
                "summary": "Eliminar receta médica",
                "operationId": "2443aab3ec5b1614ecd3282c564537cd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Receta médica eliminada"
                    },
                    "404": {
                        "description": "No encontrada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/pacientes/{paciente_id}/recetas-medicas": {
            "get": {
                "tags": [
                    "Recetas Médicas"
                ],
                "summary": "Listar recetas médicas por paciente",
                "operationId": "bba2122f3acc106fac9b8f89a58c496d",
                "parameters": [
                    {
                        "name": "paciente_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de recetas médicas del paciente"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/medicos/{medico_id}/recetas-medicas": {
            "get": {
                "tags": [
                    "Recetas Médicas"
                ],
                "summary": "Listar recetas médicas por médico",
                "operationId": "6a02c6f92259b6b937e02490acfe918b",
                "parameters": [
                    {
                        "name": "medico_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de recetas médicas del médico"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Usa el formato Bearer {token}",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}